==> 05-lights/01_geoprocess_countries.ipynb <==
import fiona
import geopandas as gpd
import glob
from rasterstats import zonal_stats
import pandas as pd
import pickle

from library.geoprocess import get_tif_ids, zonal_to_shp

shp = 'data/geo/countries/ne.shp'
gdf = gpd.GeoDataFrame.from_file(shp); gdf.head(3)

# calculate zonal stats for each country
gdf = zonal_to_shp('data/geo/images', 'data/geo/countries/countries_nf2.shp'); gdf.head(3)

# extract 1992 column
gdf_1992 = gdf[['ISO_A3', 'WB_A3', 'F101992']]; gdf_1992.head(3)

# pickle country-level zonal stats
with open('data/geo/pickles/zonal_stats_c.pickle', 'wb') as f:
    pickle.dump(gdf, f)

# pickle 1992 column
with open('data/geo/pickles/zonal_stats_c92.pickle', 'wb') as f:
    pickle.dump(gdf_1992, f)
