==> 05-lights/03_analyze_countries.ipynb <==
import geopandas as gpd
import pandas as pd
import pickle

from sklearn.linear_model import LinearRegression
import statsmodels.api as sm
import statsmodels.formula.api as smf

from library.analyze import countries_regressions

# import cleaned dataframe
df = pd.read_pickle('data/geo/pickles/zonal_stats_c_norm.pickle')

# calculate regression stats for each country
df_regression = countries_regressions(df).sort_values(by='r', ascending=False)

# save outputs to csv
df_regression.to_csv('data/country_stats.csv')

# syria case

# import cleaned dataframe
df_syria = pd.read_pickle('data/geo/pickles/zonal_stats_c_norm_syr.pickle')

# calculate regression stats for each country
df_syria_regression = countries_regressions(df_syria).sort_values(by='r', ascending=False)

# save outputs to csv
df_syria_regression.to_csv('data/country_stats_syria.csv')

# angola and south sudan cases

# import cleaned dataframe
df_agossd = pd.read_pickle('data/geo/pickles/zonal_stats_c_norm_agossd.pickle')

# calculate regression stats for each country
df_agossd_regression = countries_regressions(df_agossd).sort_values(by='r', ascending=False)

# save outputs to csv
df_agossd_regression.to_csv('data/country_stats_agossd.csv')

df_agossd_regression
